move sample from dataset serialization tests to kernel tests
